POS/requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-extensions==4.1
djangorestframework==3.16.1
drf-spectacular==0.28.0
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kombu==5.5.4
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
reportlab==4.4.3
requests==2.32.5
rpds-py==0.27.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
